pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.18.0",
    "caldav>=1.4.0",
    "python-dotenv>=1.0.1",
]
